requirements.txt
pyautogui
pillow

